refactor compile
