+++ onediff_sd_webui_extensions/compile/quantization.py
@@ -1,0 +1,28 @@
+import warnings
+from pathlib import Path
+from typing import Dict, Union
+
+from modules.sd_models import select_checkpoint
+
+from onediff.utils import logger
+
+from .utils import OneDiffCompiledGraph
+
+def get_calibrate_info(filename: str) -> Union[None, Dict]:
+    calibration_path = Path(select_checkpoint().filename).parent / filename
+    if not calibration_path.exists():
+        return None
+
+    logger.info(f"Got calibrate info at {str(calibration_path)}")
+    calibrate_info = {}
+    with open(calibration_path, "r") as f:
+        for line in f.readlines():
+            line = line.strip()
+            items = line.split(" ")
+            calibrate_info[items[0]] = [
+                float(items[1]),
+                int(items[2]),
+                [float(x) for x in items[3].split(",")],
+            ]
+    return calibrate_info
+
